[reservation] 만료된 임시예매정보 있을 경우만 제거로직 수행하도록 수정

+++ services/reservation-service/src/main/java/com/example/reservationservice/reservation/application/scheduler/DeleteTempReservationScheduler.java
@@ -24,10 +24,12 @@ public class DeleteTempReservationScheduler {
     @Transactional
     public void deleteOldTempReservations() {
         List<TempReservation> oldTempReservations = deleteTempReservationPort.getOldTempReservationsBefore(60 * 5);
-        Set<Long> seatIds = oldTempReservations.stream().map(TempReservation::seatIds).flatMap(Set::stream).collect(Collectors.toSet());
-        holdConcertSeatsPort.releaseSeats(seatIds);
-        deleteTempReservationPort.deleteOldTempReservations(
-            oldTempReservations.stream().map(TempReservation::id).toList()
-        );
+        if (!oldTempReservations.isEmpty()) {
+            Set<Long> seatIds = oldTempReservations.stream().map(TempReservation::seatIds).flatMap(Set::stream).collect(Collectors.toSet());
+            holdConcertSeatsPort.releaseSeats(seatIds);
+            deleteTempReservationPort.deleteOldTempReservations(
+                oldTempReservations.stream().map(TempReservation::id).toList()
+            );
+        }
     }
 }
